Add the possibility to install plugins
It is possible to install, update and remove official plugins using the
API as long as a security token as been generated using an admin account
and passed to ansible in the variable `sonar_token`.

Community plugins don't require this token and are installed from a user
suplied `url` key.
They can't be updated but removal is supported.